+++ android/engine/src/test/java/org/smartregister/fhircore/engine/util/extension/ResourceExtensionTest.kt
@@ -478,10 +478,14 @@ class ResourceExtensionTest : RobolectricTest() {
   @Test
   fun `Resource#generateMissingId() should generate Id if empty`() {
     val resource = Patient()
+    resource.id = "1"
 
-    Assert.assertTrue(resource.logicalId.isEmpty())
     resource.generateMissingId()
+    Assert.assertEquals("1", resource.logicalId)
 
+    resource.id = null
+    resource.generateMissingId()
+    Assert.assertNotEquals("1", resource.logicalId)
     Assert.assertFalse(resource.logicalId.isEmpty())
   }
 
@@ -559,5 +563,42 @@ class ResourceExtensionTest : RobolectricTest() {
     Assert.assertTrue(questionnaire[1].readOnly)
     Assert.assertTrue(questionnaire[2].hasExtension(FhirCoreQuestionnaireFragment.BARCODE_URL))
     Assert.assertTrue(questionnaire[2].readOnly)
+  }
+
+  @Test
+  fun `Type valueToString() should return string representation`() {
+    Assert.assertEquals("12345", StringType("12345").valueToString())
+    Assert.assertEquals("true", BooleanType(true).valueToString())
+    Assert.assertEquals(Date().makeItReadable(), DateTimeType(Date()).valueToString())
+    Assert.assertEquals("d", Coding("s", "c", "d").valueToString())
+    Assert.assertEquals(
+      "d",
+      CodeableConcept().apply { addCoding(Coding("s", "c", "d")) }.valueToString()
+    )
+    Assert.assertEquals(
+      "3.4",
+      Quantity()
+        .apply {
+          this.value = BigDecimal.valueOf(3.4)
+          this.unit = "G"
+        }
+        .valueToString()
+    )
+  }
+
+  @Test
+  fun `Resource generateReferenceValue() should return correct reference`() {
+    val resource = Patient().apply { id = "123456" }
+
+    Assert.assertEquals("Patient/123456", resource.referenceValue())
+  }
+
+  @Test
+  fun `Resource#isPatient() should return true if resource is valid`() {
+    val resource = Patient()
+    Assert.assertFalse(resource.isPatient("1"))
+
+    resource.id = "1"
+    Assert.assertTrue(resource.isPatient("1"))
   }
 }