+++ CHANGELOG.md
@@ -37,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Quest | Added G6PD login configuration
 - Quest | Add Swahili translations for patient registration questionnaire and activate language switcher
 - Quest | Swahili localization
+- Quest | Added G6PD Patient Details page to show G6PD Status and Test Results
 
 - EIR | Added workflow configurations
 - EIR | Add Adverse Events feature to log any reactions after the first dose of vaccine using Structure map